refactor: move io class file to io folder
