+++ src/components/Lotto.js
@@ -1,6 +1,6 @@
 import { DELIMETER, InputMessages } from '../resources/Constants.js';
 import winningNumberValidator from '../utils/validation/winningNumberValidator.js';
-import Input from '../utils/Input.js';
+import Input from '../utils/io/Input.js';
 
 class Lotto {
   #numbers;
